First Commit
I've assigned the four faders for 4 different channels. The master volume is assigned to the fourth endless encoder (the most right encoder on the top row). Finally, I started attempting to map transport control.  As a demonstration, the play button is mapped to the button that sends out the midi note 36 (button with an A written on top)